+++ server/monitor-rest/src/main/java/org/apache/accumulo/monitor/rest/api/LogEvent.java
@@ -16,7 +16,7 @@
  */
 package org.apache.accumulo.monitor.rest.api;
 
-import org.codehaus.jackson.annotate.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
  * A single message logged
